+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog]
   - Add a note if DQT active alert flag is set to true
   - Add important notes to claim summary (DQT active alert)
   - Add ECP admin claim matching details task screen
+  - DQT identity check mailer removed
 - All policies
   - Home Address selection
     - Add home address search screen to claim journey